Conflicts:
	apps/comment/service/business.py
	apps/socialactions/service/business.py
	apps/userprofile/views.py
	templates/article/single.html

+++ apps/userprofile/service/business.py
@@ -2,8 +2,9 @@ from django.contrib.auth import authenticate
 from django.contrib.auth.models import User
 from django.utils.translation import ugettext as _
 from django.db import transaction
+from apps.taxonomy.models import Term, Taxonomy
 
-from apps.userprofile.models import UserProfile, Country, State, City, Occupation
+from apps.userprofile.models import UserProfile, Country, State, City, Occupation, Responsibility
 
 
 def check_user_exists(username_or_email=None):
@@ -51,8 +52,7 @@ def create_profile(user, data=None):
 def edit_profile(user, data_profile=None, data_formset=None):
 
     try:
-        if data_profile:
-            profile = update_profile(user, data_profile)
+        profile = update_profile(user, data_profile)
 
         if data_formset:
             for data in data_formset:
@@ -67,31 +67,54 @@ def update_profile(user=None, data=None):
 
     profile = get_profile(user)
 
+    if data:
+        try:
+            profile.birth = data['birth']
+            profile.gender = data['gender']
+            profile.city = data['city']
+
+            if data['profile_picture'] and profile.profile_picture != data['profile_picture']:
+                profile.profile_picture.delete()
+                profile.profile_picture = data['profile_picture']
+
+            profile.save()
+        except:
+            return False
+
+    return profile
+
+def update_user(user, data={}):
+
     try:
-        profile.birth = data['birth']
-        profile.gender = data['gender']
-        profile.city = data['city']
-
-        if data['profile_picture'] and profile.profile_picture != data['profile_picture']:
-            profile.profile_picture.delete()
-            profile.profile_picture = data['profile_picture']
-
-        profile.save()
+        user = User.objects.filter(id=user.id).update(**data)
     except:
         return False
+
+    return user
+
+
+def update_wizard_step(profile, step=0):
+
+    try:
+        profile.wizard_step = step
+        profile.save()
+    except:
+        return None
 
     return profile
 
 
-def create_occupation(user=None, data={}):
+def create_occupation(profile=None, user=None, data={}):
 
-    profile = get_profile(user)
+    if profile:
+        profile = profile
+    elif user:
+        profile = get_profile(user)
+
+    data.update({'profile': profile})
 
     try:
-        occupation = Occupation()
-        occupation.responsibility = data['responsibility']
-        occupation.company = data['company']
-        occupation.profile = profile
+        occupation = Occupation(**data)
         occupation.save()
     except:
         return False
@@ -159,3 +182,20 @@ def update_occupation(occupation, data):
         return False
 
     return occupation
+
+
+def get_responsibilities():
+
+    try:
+        responsibilities = Responsibility.objects.all()
+    except Responsibility.DoesNotExist:
+        return False
+
+    return responsibilities
+
+
+def get_categories():
+    term = Term.objects.filter(description="Categoria")
+    categories = Taxonomy.objects.filter(term=term)
+
+    return categories
